fix(ci): scale paginated runtime exports
